requirements.txt
opencv-python>=4.8.0
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.0.0
mediapipe>=0.10.1
flask>=2.3.2
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
